requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
celery==5.3.1
redis==4.6.0
psycopg2-binary==2.9.7
google-api-python-client==2.95.0
google-auth==2.22.0
requests==2.31.0
beautifulsoup4==4.12.2
langdetect==1.0.9
python-dotenv==1.0.0
pydantic==2.1.1
SQLAlchemy==2.0.19
alembic==1.11.1
gunicorn==21.2.0
mysql-connector-python==8.1.0